Fix webhook error (#24)
Webhook assignment was happening outside the Lambda flow, which makes no sense. If run from CLI this environment variable will not be set, and we would probably not want a Teams message anyway. Relocated to Lambda handler and updated Teams function to expect the webhook to be passed as a parameter.