CloudWatch logs, Terraform (#55)
Edited the Terraform code to make sure that CloudWatch logs are set up correctly for both Lambda functions.

While in there, I took the chance to tidy up a few odds and ends around names. In so doing, I eliminated some of the variables. We just use the Lambda function names for everything else (logs, SQS).